more merging

+++ nDPId_pcap.c
@@ -3689,7 +3689,7 @@ static void write_to_socket(struct nDPId_reader_thread * const reader_thread,
     }
 
     char * converted_json_str = NULL;
-    size_t flow_risk_count = 0;
+    int flow_risk_count = 0;
 
     ConvertnDPIDataFormat(json_msg, json_string_with_http_or_tls_info, 0, &converted_json_str, &flow_risk_count);
     if (converted_json_str != NULL)
@@ -3700,10 +3700,10 @@ static void write_to_socket(struct nDPId_reader_thread * const reader_thread,
             if (flow_risk_count)
             {
                 write_to_socket_2(reader_thread, converted_json_str, length);
-                for (size_t index = 1; index < flow_risk_count; index++)
+                for (int index = 1; index < flow_risk_count; index++)
                 {
                     free(converted_json_str);
-                    size_t flow_risk_count_dummy = 0;
+                    int flow_risk_count_dummy = 0;
                     ConvertnDPIDataFormat(json_msg,
                                           json_string_with_http_or_tls_info,
                                           index,
